requirements.txt
python-dotenv
pyTelegramBotAPI
reportlab